prospector/requirements.txt
googlemaps>=4.10.0
playwright>=1.47.0
beautifulsoup4>=4.12.3
requests>=2.32.3
colorthief>=0.2.1
Pillow>=10.4.0
scikit-image>=0.24.0
scikit-learn>=1.5.2
ollama>=0.3.3
jinja2>=3.1.4
python-dotenv>=1.0.1
rich>=13.9.2
